requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
numpy==1.18.5
pandas==1.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
